refactor: extract useSpellAudio hook from App.tsx
